fix: AuthGate types and refactor
